refactor playback and add scenario
