+++ src/test/perf/verify.ts
@@ -1,9 +1,7 @@
-import { JSONRPCError } from 'jayson/promise';
-import { Result } from 'neverthrow';
 import { RPCClient } from '~/network/rpc';
-import { Message } from '~/types';
 import { sleep } from '~/utils/crypto';
 import { logger } from '~/utils/logger';
+import { ComputeSetDifference } from '~/test/perf/utils';
 
 // A 5 minute network sync timer
 let syncTimer: NodeJS.Timer;
@@ -31,28 +29,19 @@ export const waitForSync = async (rpcClients: RPCClient[]) => {
   clearInterval(syncTimer);
 };
 
-const compareMessages = (
-  left: Result<Set<number | Message>, JSONRPCError>,
-  right: Result<Set<number | Message>, JSONRPCError>
-) => {
-  // This is hacky and not going to work for very long (FlatBuffs?)
-  // TODO Consider pulling in lodash for dev deps
-  if (left.isErr() || right.isErr()) return false;
-  return JSON.stringify(left.value) === JSON.stringify(right.value);
-};
-
 /** Caches information from 1 RPC Client (Hub) and compares it to a list of other RPC Clients (Hubs)  */
 const cacheCompareClients = async (sourceRpcClient: RPCClient, networkRpcClients: RPCClient[]) => {
   // Check that the hubs have synchronized
   const userIds = await sourceRpcClient.getUsers();
   if (userIds.isErr()) {
-    logger.error(userIds.error, `Unable to get users from ${sourceRpcClient.getServerMultiaddr()}`);
+    logger.error(userIds.error, `Unable to get users from ${sourceRpcClient.serverMultiaddr}`);
     return false;
   }
   for (const client of networkRpcClients) {
     const clientSet = await client.getUsers();
-    if (!compareMessages(clientSet, userIds)) {
-      logger.error(`${client.getServerMultiaddr()} fids are not in sync`);
+    const difference = ComputeSetDifference(userIds, clientSet);
+    if (difference.isErr() || (difference.isOk() && difference.value.size > 0)) {
+      logger.error(difference, `${client.serverMultiaddr} fids are not in sync`);
       return false;
     }
   }
@@ -60,32 +49,37 @@ const cacheCompareClients = async (sourceRpcClient: RPCClient, networkRpcClients
     const casts = await sourceRpcClient.getAllCastsByUser(user);
     for (const client of networkRpcClients) {
       const clientSet = await client.getAllCastsByUser(user);
-      if (!compareMessages(clientSet, casts)) {
-        logger.error(`${client.getServerMultiaddr()} casts are not in sync for user ${user}`);
+      const difference = ComputeSetDifference(casts, clientSet);
+      if (difference.isErr() || (difference.isOk() && difference.value.size > 0)) {
+        logger.error(difference, `${client.serverMultiaddr} casts are not in sync for user ${user}`);
         return false;
       }
     }
     const follows = await sourceRpcClient.getAllFollowsByUser(user);
     for (const client of networkRpcClients) {
       const clientSet = await client.getAllFollowsByUser(user);
-      if (!compareMessages(clientSet, follows)) {
-        logger.error(`${client.getServerMultiaddr()} follows are not in sync for user ${user}`);
+      const difference = ComputeSetDifference(follows, clientSet);
+      if (difference.isErr() || (difference.isOk() && difference.value.size > 0)) {
+        logger.error(difference, `${client.serverMultiaddr} follows are not in sync for user ${user}`);
+
         return false;
       }
     }
     const reactions = await sourceRpcClient.getAllReactionsByUser(user);
     for (const client of networkRpcClients) {
       const clientSet = await client.getAllReactionsByUser(user);
-      if (!compareMessages(clientSet, reactions)) {
-        logger.error(`${client.getServerMultiaddr()} reactions are not in sync for user ${user}`);
+      const difference = ComputeSetDifference(reactions, clientSet);
+      if (difference.isErr() || (difference.isOk() && difference.value.size > 0)) {
+        logger.error(difference, `${client.serverMultiaddr} reactions are not in sync for user ${user}`);
         return false;
       }
     }
     const verifications = await sourceRpcClient.getAllVerificationsByUser(user);
     for (const client of networkRpcClients) {
       const clientSet = await client.getAllVerificationsByUser(user);
-      if (!compareMessages(clientSet, verifications)) {
-        logger.error(`${client.getServerMultiaddr()} verifications are not in sync for user ${user}`);
+      const difference = ComputeSetDifference(verifications, clientSet);
+      if (difference.isErr() || (difference.isOk() && difference.value.size > 0)) {
+        logger.error(difference, `${client.serverMultiaddr} verifications are not in sync for user ${user}`);
         return false;
       }
     }
